moved anything that depends on external dependences in io/event.js
