+++ src/store/event/eventReducer.js
@@ -19,9 +19,9 @@ import {
   updatePublicEvent,
   removePublicEvent,
   addPublicEvent,
-  uuid,
   createSessionChat
-} from "./eventAction";
+} from "../../io/event";
+import { uuid } from "../../io/eventFN";
 
 let initialState = {
   allEvents: [],
@@ -43,26 +43,26 @@ export default function reduce(state = initialState, action = {}) {
     case VIEW_EVENT:
       return { ...state, currentEvent: action.payload.eventInfo };
     case REMOVE_EVENT:
-      var newState = state;
-      newState.allEvents = newState.allEvents.filter(function(obj) {
+      let { allEvents } = state;
+      allEvents = allEvents.filter(function(obj) {
         return obj && obj.id !== action.payload.obj.id;
       });
       publishEvents(action.payload.obj.uid, removePublicEvent);
-      saveEvents("default", newState.allEvents);
-      return newState;
+      saveEvents("default", allEvents);
+      return { ...state, allEvents };
     case ADD_EVENT:
-      var newState2 = state;
+      var { allEvents } = state;
       action.payload.calendarName = "default";
-      newState2.allEvents.push(action.payload);
+      allEvents.push(action.payload);
       if (action.payload.public) {
         publishEvents(action.payload, addPublicEvent);
       }
-      saveEvents("default", newState2.allEvents);
-      return newState2;
+      saveEvents("default", allEvents);
+      return { ...state, allEvents };
     case UPDATE_EVENT:
-      var newState3 = state;
+      var { allEvents } = state;
       var eventInfo = action.payload.obj;
-      newState3.allEvents[eventInfo.id] = eventInfo;
+      allEvents[eventInfo.id] = eventInfo;
       if (eventInfo.public) {
         eventInfo.uid = eventInfo.uid || uuid();
         publishEvents(eventInfo, updatePublicEvent);
@@ -71,10 +71,10 @@ export default function reduce(state = initialState, action = {}) {
           publishEvents(eventInfo.uid, removePublicEvent);
         }
       }
-      saveEvents("default", newState3.allEvents);
-      return newState3;
+      saveEvents("default", allEvents);
+      return { ...state, allEvents };
     case INVITES_SENT:
-      var allEvents = state.allEvents;
+      var { allEvents } = state;
       if (action.payload.type === "add") {
         allEvents.push(action.payload.eventInfo);
         saveEvents("default", allEvents);
